+++ projects/ngx-location-picker/src/lib/components/ngx-location-picker.component.ts
@@ -133,6 +133,8 @@ export class NgxLocationPickerComponent implements OnInit, OnDestroy, ControlVal
   @Input() filterLayers: FilterLayerOptions[];
   /* If provided, adds coordinate to resultList at index */
   @Input() addCoordinateToResultsAt?: number = null;
+  /* If search string contains one of these words, search for locations instead of address */
+  @Input() locationKeywords: string[] = ['kaainummer'];
   /* AddPolygon event */
   @Output() addPolygon = new EventEmitter<any>();
   /* AddLine event */
@@ -456,7 +458,8 @@ export class NgxLocationPickerComponent implements OnInit, OnDestroy, ControlVal
         cascadingCoordinateReturnSingle: this.cascadingCoordinateReturnSingle,
         cascadingCoordinateLimit: this.cascadingCoordinateLimit,
         cascadingCoordinateRules: this.cascadingCoordinateRules,
-        selectedLocation: this.previousLocation
+        selectedLocation: this.previousLocation,
+        locationKeywords: this.locationKeywords
       };
 
       this.locationServiceSubscription = this.locationPickerService.delegateSearch(
@@ -465,7 +468,7 @@ export class NgxLocationPickerComponent implements OnInit, OnDestroy, ControlVal
         this.foundLocations = response;
 
         //adds used coordinate to result list
-        if (this.addCoordinateToResultsAt) {
+        if (this.addCoordinateToResultsAt && this.pickedLocation) {
           this.foundLocations.splice(this.addCoordinateToResultsAt, 0, this.previousLocation);
         }