+++ CHANGELOG.md
@@ -20,6 +20,13 @@ All notable changes to this project will be documented in this file.
     - Removed code that already exists in ngx-location-viewer (zoomin, zoomout & tilelayers)
     - Implemented all locationviewer params
 
+## [4.8.1] - 2021-05-10
+
+### Fixed
+- Added locationKeywords param with default value ['kaainummer'], if 'kaainummer number' is entered search for location instead of address
+- When param addCoordinateToResultsAt is provided: only add result when search is triggered by coordinate
+
+
 ## [4.8.0] - 2021-04-19
 
 ### Fixed
@@ -158,7 +165,8 @@ All notable changes to this project will be documented in this file.
 
 - Initial release.
 
-[Unreleased]: https://github.com/digipolisantwerp/location-picker_widget_angular/compare/v4.8.0...HEAD
+[Unreleased]: https://github.com/digipolisantwerp/location-picker_widget_angular/compare/v4.8.1...HEAD
+[4.8.1]: https://github.com/digipolisantwerp/location-picker_widget_angular/compare/v4.8.0...v4.8.1
 [4.8.0]: https://github.com/digipolisantwerp/location-picker_widget_angular/compare/v4.7.1...v4.8.0
 [4.7.1]: https://github.com/digipolisantwerp/location-picker_widget_angular/compare/v4.7.0...v4.7.1
 [4.7.0]: https://github.com/digipolisantwerp/location-picker_widget_angular/compare/v4.6.0...v4.7.0